chore: Update babel.config.js and index.js
The changes in babel.config.js update the presets array to use "@react-native/babel-preset" for compiling React Native code. The addition of a comment explains the purpose of the file.

In index.js, a comment is added to explain the purpose of the file and its relation to package.json.

```